.Net: Restructure function calling samples (#6620)
This PR renames the `AutoFunctionCalling` folder to `FunctionCalling` to
reflect that the samples it contains may demonstrate not only auto
function calling but also manual function calling. Additionally, this PR
adds XML comments to the sample classes and their public methods.
